requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.111.0
uvicorn==0.32.1
python-multipart==0.0.19
starlette==0.37.2

# Database dependencies
pymongo==4.9.0
motor==3.6.0

# Authentication and security
PyJWT==2.10.1
bcrypt==4.2.1
python-jose==3.3.0


# Azure Blob Storage
azure-storage-blob==12.24.0
azure-core==1.32.0

# Data validation and serialization
pydantic==2.10.3
pydantic-settings==2.9.1
email-validator==2.2.0

# Environment and configuration
python-dotenv==1.0.1

# HTTP client
httpx==0.28.0
requests==2.32.3

# Utilities
python-dateutil==2.9.0.post0